CNDB-9974: Add test that a node cannot start with enable_scripted_user_defined_functions=true
